backend/requirements.txt
# CozyClean Backend Dependencies
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# ORM & database driver
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Validation & config
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Rate limiting
slowapi>=0.1.9

# JWT authentication
python-jose[cryptography]>=3.3.0
